datagen/requirements.txt
fastapi==0.100.0
starlette==0.27.0
httpx==0.24.1
uvicorn==0.22.0
pydantic==1.10.7
pandas
python-dotenv
requests
python-multipart
ifcopenshell
pytest